requirements.txt
streamlit==1.44.1,
torch==2.7.1+cu128,
torchvision==0.22.1+cu128,
numpy==2.2.4,
pillow==11.2.1,
matplotlib==3.10.3
scikit-learn==1.6.1